builtin-types-part2 (#13)

+++ docs/examples/userguide/wrapping_CPlusPlus/templates.pyx
@@ -12,17 +12,17 @@ cdef extern from "<vector>" namespace "std":
             bint operator!=(iterator)
         vector()
         void push_back(T&)
-        T& operator[](int)
-        T& at(int)
+        T& operator[](usize)
+        T& at(usize)
         iterator begin()
         iterator end()
 
-cdef vector[int] *v = new vector[int]()
+cdef vector[i32] *v = new vector[i32]()
 cdef int i
 for i in range(10):
     v.push_back(i)
 
-cdef vector[int].iterator it = v.begin()
+cdef vector[i32].iterator it = v.begin()
 while it != v.end():
     print(deref(it))
     inc(it)
